Try to fix MSVC compiler error
